% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evaluate_config.R
\name{eval_config}
\alias{eval_config}
\title{Run an algorithm}
\usage{
eval_config(
  learner,
  task_id,
  configuration,
  trainsize,
  logfile = NULL,
  seed = NULL
)
}
\arguments{
\item{learner}{\code{character}\cr
Learner name e.g. "classif.rpart".}

\item{task_id}{\code{character}\cr
OpenML task ID.}

\item{configuration}{\code{list}\cr
Named list of hyperparameters.}

\item{trainsize}{\code{numeric}\cr
Fraction of training set size.}

\item{logfile}{\code{character}\cr
Logfile to write to.}

\item{seed}{\code{character}\cr
Seed to set. Defaults to \code{NULL} (no custom seed, using original one).}
}
\description{
Run an algorithm.
}
\examples{
learner = "classif.rpart"
task_id = "190424"
configuration = list(learner.classif.rpart.cp = 0.5,
  learner.classif.rpart.maxdepth = 7,
  learner.classif.rpart.minbucket = 50,
  learner.classif.rpart.minsplit = 20,
  learner.branch.selection = "reweighing_os",
  learner.reweighing_os.alpha = 0.05,
  branch.selection = "nop")
trainsize = 0.11
eval_config(learner, task_id, configuration, trainsize)
}
