% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evaluate_config.R
\name{eval_fair}
\alias{eval_fair}
\title{Run an algorithm from fair}
\usage{
eval_fair(learner, task_id, configuration, ...)
}
\arguments{
\item{learner}{\code{character}\cr
Learner name e.g. "fair_rpart".}

\item{task_id}{\code{character}\cr
OpenML task ID.}

\item{configuration}{\code{list}\cr
Named list of hyperparameters.}

\item{...}{\code{any}\cr
Arguments passed on to \code{eval_config}.}
}
\description{
Run an algorithm from fair.
}
\examples{
learner = "fair_rpart"
task_id = "190424"
configuration = list(cp = 0.5, maxdepth = 7, minbucket = 50, minsplit = 20,
  pre_post = "pre", reweighing_os_alpha = 0.05, trainsize = 0.11)
eval_fair(learner, task_id, configuration)
}
