\name{FAiR-internal}
\alias{FAiR-internal}

\alias{make_parameter,parameter.cormat-method}
\alias{make_parameter,parameter.scale-method}
\alias{make_parameter,parameter.coef-method}
\alias{make_parameter,parameter.coef.nl-method}
\alias{make_parameter,parameter.coef.SEFA-method}
\alias{make_parameter,parameter.coef.SEFA.nl-method}

\alias{restrictions2model,restrictions.orthonormal,manifest.basic-method}
\alias{restrictions2model,restrictions.independent,manifest.basic-method}
\alias{restrictions2model,restrictions.1storder,manifest.basic-method}
\alias{restrictions2model,restrictions.1storder.EFA,manifest.basic-method}
\alias{restrictions2model,restrictions.general,manifest.basic-method}
\alias{restrictions2model,restrictions.2ndorder,manifest.basic-method}
\alias{fitS4,restrictions.independent,manifest.basic-method}
\alias{fitS4,restrictions.factanal,manifest.basic-method}
\alias{gr_fitS4,restrictions.factanal,manifest.basic-method}
\alias{gr_fitS4,restrictions.orthonormal,manifest.basic-method}
\alias{gr_fitS4,restrictions.1storder,manifest.basic-method}
\alias{gr_fitS4,restrictions.general,manifest.basic-method}
\alias{gr_fitS4,restrictions.2ndorder,manifest.basic-method}
\alias{bfgs_helpS4,restrictions.factanal,manifest.basic-method}
\alias{create_start,restrictions.factanal,manifest.basic-method}
\alias{create_start,restrictions.orthonormal,manifest.basic-method}
\alias{create_start,restrictions.1storder,manifest.basic-method}
\alias{create_start,restrictions.general,manifest.basic-method}
\alias{create_start,restrictions.2ndorder,manifest.basic-method}
\alias{create_FAobject,restrictions.independent,manifest.basic-method}
\alias{create_FAobject,restrictions.factanal,manifest.basic-method}
\alias{create_FAobject,restrictions.orthonormal,manifest.basic-method}
\alias{create_FAobject,restrictions.1storder,manifest.basic-method}
\alias{create_FAobject,restrictions.general,manifest.basic-method}
\alias{create_FAobject,restrictions.2ndorder,manifest.basic-method}
\alias{restrictions2draws,restrictions.orthonormal,manifest.basic-method}
\alias{restrictions2draws,restrictions.1storder,manifest.basic-method}
\alias{restrictions2draws,restrictions.1storder.EFA,manifest.basic-method}
\alias{restrictions2draws,restrictions.general,manifest.basic-method}
\alias{restrictions2draws,restrictions.2ndorder,manifest.basic-method}

\alias{coef,FA.general-method}
\alias{coef,FA.2ndorder-method}
\alias{coef,restrictions.independent-method}
\alias{coef,restrictions.factanal-method}
\alias{coef,restrictions.orthonormal-method}
\alias{coef,restrictions.1storder-method}
\alias{coef,restrictions.general-method}
\alias{coef,restrictions.2ndorder-method}

\alias{cormat,restrictions.factanal-method}
\alias{cormat,restrictions.orthonormal-method}

\alias{fitted,restrictions.independent-method}
\alias{fitted,restrictions.factanal-method}
\alias{fitted,restrictions.orthonormal-method}

\alias{loadings,restrictions.factanal-method}

\alias{plot,FA.general,ANY-method}
\alias{plot,FA.2ndorder,ANY-method}
\alias{plot,manifest.data.mcd,ANY-method}

\alias{profile,FA.general-method}
\alias{profile,FA.2ndorder-method}

\alias{restrictions2RAM,restrictions.orthonormal-method}
\alias{restrictions2RAM,restrictions.1storder-method}
\alias{restrictions2RAM,restrictions.general-method}
\alias{restrictions2RAM,restrictions.2ndorder-method}

\alias{screeplot,manifest.data-method}

\alias{show,manifest.data-method}
\alias{show,manifest.data.ordinal-method}
\alias{show,manifest.data.ranks-method}
\alias{show,restrictions.factanal-method}
\alias{show,restrictions.orthonormal-method}
\alias{show,restrictions.1storder-method}
\alias{show,restrictions.1storder.EFA-method}
\alias{show,restrictions.general-method}
\alias{show,restrictions.2ndorder-method}

\alias{uniquenesses,restrictions.factanal-method}

\title{Quasi-internal S4 Methods}
\description{
  This page \dQuote{documents} various things to prevent them
  from cluttering the index to the documentation. In virtually all
  cases, users would not be interested in them, whereas developers
  would be more interested in the source code. If you landed here accidentally,
  try a less specific search or see the links in the See Also section below.
}
\seealso{\code{\link{fitS4-methods}},
         \code{\link{bfgs_fitS4-methods}},
         \code{\link{gr_fitS4-methods}},
         \code{\link{bfgs_helpS4-methods}},
         \code{\link{create_start-methods}},
         \code{\link{create_FAobject-methods}},
         \code{\link{restrictions2draws-methods}},
         \code{\link{S3methodsFAiR}}
}
\author{Ben Goodrich}
\keyword{internal}
