% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\docType{package}
\name{fair-package}
\alias{fair}
\alias{fair-package}
\title{fair: Reproduce results from fair}
\description{
Reproduce results from fair including seeds.
}
\author{
\strong{Maintainer}: Lennart Schneider \email{lennart.sch@web.de} (\href{https://orcid.org/0000-0003-4152-5308}{ORCID})

}
