% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rescale.R
\name{rescale}
\alias{rescale}
\title{Rescaling vectors}
\usage{
rescale(v)
}
\arguments{
\item{v}{numeric vectors}
}
\value{
numeric vecotr, rescaled from input vector to the range of 0-1
}
\description{
Rescaling vectors
}
\examples{
rescale(c(1,2,3)) --> 0.0 0.5 1.0
rescale(c(1,2,3,4,5)) --> 0.00 0.25 0.50 0.75 1.00
}
