% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/center.R
\name{center}
\alias{center}
\title{Centering Datasets}
\usage{
center(data, desired)
}
\arguments{
\item{data}{the numeric vector to be centered}

\item{desired}{numeric mid-point value around which the data will be centered}
}
\value{
this function returns a new vector containing the original _data_ centered around the _desired_ values
}
\description{
Centering Datasets
}
\examples{
center(c(1,2,3),0) -> -1 0 1
}
