% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rescale.R
\name{rescale}
\alias{rescale}
\title{Rescaling vectors}
\usage{
rescale(v, upper = 1, lower = 0)
}
\arguments{
\item{v}{numeric vectors}

\item{upper}{numeric value of upper rescaling value default: 1}

\item{lower}{numeric value of lower rescaling value default: 0}
}
\value{
numeric vecotr, rescaled from input vector to the range of 0-1
}
\description{
Rescaling vectors
}
\examples{
rescale(c(1,2,3)) --> 0.0 0.5 1.0
rescale(c(1,2,3,4,5)) --> 0.00 0.25 0.50 0.75 1.00
rescale(c(1,2,3), 1,2) --> 1.0 1.5 2.0
}
